Moving to structs and defaults file.
