--- crates/utils/src/settings/rate_limit_config.rs
@@ -1,28 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct RateLimitConfig {
-  pub message: i32,
-  pub message_per_second: i32,
-  pub post: i32,
-  pub post_per_second: i32,
-  pub register: i32,
-  pub register_per_second: i32,
-  pub image: i32,
-  pub image_per_second: i32,
-}
-
-impl Default for RateLimitConfig {
-  fn default() -> Self {
-    Self {
-      message: 180,
-      message_per_second: 60,
-      post: 6,
-      post_per_second: 600,
-      register: 3,
-      register_per_second: 3600,
-      image: 6,
-      image_per_second: 3600,
-    }
-  }
-}
